requirements.txt
Django>=6.0
djangorestframework>=3.14
django-cors-headers>=4.0
drf-spectacular>=0.27
openpyxl>=3.1
Pillow>=10.0
qrcode>=7.4
django-filter>=25.2
reportlab>=4.0
